tweak: moved stuff from References.cpp
