+++ Code/client/Games/Overrides.cpp
@@ -2,6 +2,7 @@
 
 #include <Games/Overrides.h>
 
+thread_local uint32_t ScopedReferencesOverride::s_refCount = 0;
 thread_local uint32_t ScopedEquipOverride::s_refCount = 0;
 thread_local uint32_t ScopedUnequipOverride::s_refCount = 0;
 thread_local uint32_t ScopedSaveLoadOverride::s_refCount = 0;
