+++ Code/client/Games/Skyrim/AI/AIProcess.cpp
@@ -1,1 +1,8 @@
 #include "AIProcess.h"
+
+void AIProcess::KnockExplosion(Actor* apActor, const NiPoint3* aSourceLocation, float afMagnitude)
+{
+    TP_THIS_FUNCTION(TKnockExplosion, void, AIProcess, Actor*, const NiPoint3*, float);
+    POINTER_SKYRIMSE(TKnockExplosion, knockExplosion, 39895);
+    TiltedPhoques::ThisCall(knockExplosion, this, apActor, aSourceLocation, afMagnitude);
+}
