+++ Code/client/Services/Generic/QuestService.cpp
@@ -23,7 +23,6 @@ static TESQuest* FindQuestByNameId(const String& name)
     return it != questRegistry.end() ? *it : nullptr;
 }
 
-// TODO: ft (verify)
 QuestService::QuestService(World& aWorld, entt::dispatcher& aDispatcher)
     : m_world(aWorld)
 {
@@ -169,7 +168,6 @@ bool QuestService::StopQuest(uint32_t aformId)
     return false;
 }
 
-// TODO: ft (verify)
 bool QuestService::IsNonSyncableQuest(TESQuest* apQuest)
 {
     // non story quests are "blocked" and not synced
